Melhorias para entrega de nova versão

+++ web/lib/WebZeta/Controller/Admin/Form/Candidate.pm
@@ -35,18 +35,20 @@ sub process : Chained('base') : PathPart('candidate') : Args(0) {
 		if( $c->req->upload ) {
 			my $upload = $c->req->upload('img_profile');
 			
-			$api->stash_result(
-				$c, 'candidates/upload_file',
-				method => 'UPLOAD',
-				body   => [
-					candidate_id 	=> $c->stash->{id},
-					file 			=> [ $upload->tempname ],
-					type			=> 'profile'
-				]
-			);
-		
-			if ( $c->stash->{error} ) {
-				$c->detach( '/form/redirect_error', [ '/admin/candidate/index', {}, 'Problemas ao associar imagem de perfil ao candidato.' ] );
+			if($upload) {
+				$api->stash_result(
+					$c, 'candidates/upload_file',
+					method => 'UPLOAD',
+					body   => [
+						candidate_id 	=> $c->stash->{id},
+						file 			=> [ $upload->tempname ],
+						type			=> 'profile'
+					]
+				);
+			
+				if ( $c->stash->{error} ) {
+					$c->detach( '/form/redirect_error', [ '/admin/candidate/index', {}, 'Problemas ao associar imagem de perfil ao candidato.' ] );
+				}
 			}
 			
 			$upload = $c->req->upload('gvt_program');
@@ -61,10 +63,10 @@ sub process : Chained('base') : PathPart('candidate') : Args(0) {
 						type			=> 'program'
 					]
 				);
-			}
 		
-			if ( $c->stash->{error} ) {
-				$c->detach( '/form/redirect_error', [ '/admin/candidate/index', {}, 'Problemas ao cadastrar programa de governo ao candidato.' ] );
+				if ( $c->stash->{error} ) {
+					$c->detach( '/form/redirect_error', [ '/admin/candidate/index', {}, 'Problemas ao cadastrar programa de governo ao candidato.' ] );
+				}
 			}
 		}
